requirements.txt
aiogram==3.20.0.post0
beautifulsoup4==4.13.4
lxml==5.4.0
openpyxl==3.1.5
pandas==2.2.3
python-dotenv==1.1.0
requests==2.32.3
